requirements.txt
alembic==1.8.1
fastapi==0.94.1
fastapi_amis_admin==0.5.0
fastapi_mail==1.2.6
passlib==1.7.4
pycryptodome==3.17
pydantic==1.10.2
PyJWT==2.6.0
pyotp==2.6.0
python-dotenv==1.0.0
SQLAlchemy==1.4.41
uvicorn==0.21.0
psycopg2-binary==2.9.5
asyncpg==0.27.0
requests==2.28.2
boto3==1.26.115